+++ crates/icp-cli/src/commands/cycles/transfer.rs
@@ -1,12 +1,14 @@
-use candid::Principal;
+use anyhow::ensure;
 use clap::Args;
 use icp::context::Context;
 use icp_canister_interfaces::cycles_ledger::{CYCLES_LEDGER_BLOCK_FEE, CYCLES_LEDGER_PRINCIPAL};
+use icrc_ledger_types::icrc1::account::Account;
 
 use crate::commands::args::TokenCommandArgs;
-use crate::commands::parsers::parse_cycles_amount;
+use crate::commands::parsers::{parse_cycles_amount, parse_subaccount};
 use crate::operations::token::transfer::icrc1_transfer;
 
+/// Transfer cycles to another principal
 #[derive(Debug, Args)]
 pub(crate) struct TransferArgs {
     /// Cycles amount to transfer.
@@ -15,13 +17,29 @@ pub(crate) struct TransferArgs {
     pub(crate) amount: u128,
 
     /// The receiver of the cycles transfer
-    pub(crate) receiver: Principal,
+    pub(crate) receiver: Account,
+
+    /// The subaccount to transfer to (only if the receiver is a principal)
+    #[arg(long, value_parser = parse_subaccount)]
+    pub(crate) to_subaccount: Option<[u8; 32]>,
+
+    //// The subaccount to transfer from
+    #[arg(long, value_parser = parse_subaccount)]
+    pub(crate) from_subaccount: Option<[u8; 32]>,
 
     #[command(flatten)]
     pub(crate) token_command_args: TokenCommandArgs,
 }
 
 pub(crate) async fn exec(ctx: &Context, args: &TransferArgs) -> Result<(), anyhow::Error> {
+    ensure!(
+        !(args.to_subaccount.is_some() && args.receiver.subaccount.is_some()),
+        "Cannot use both --subaccount with an account ID. Use a plain principal if you want to change the subaccount."
+    );
+    let mut receiver = args.receiver;
+    if let Some(subaccount) = args.to_subaccount {
+        receiver.subaccount = Some(subaccount);
+    }
     let selections = args.token_command_args.selections();
 
     // Agent
@@ -36,9 +54,10 @@ pub(crate) async fn exec(ctx: &Context, args: &TransferArgs) -> Result<(), anyho
     // Execute transfer
     let transfer_info = icrc1_transfer(
         &agent,
+        args.from_subaccount,
         CYCLES_LEDGER_PRINCIPAL,
         args.amount.into(),
-        args.receiver,
+        receiver,
         CYCLES_LEDGER_BLOCK_FEE.into(),
         0,
         "cycles".to_string(),