+++ crates/icp-cli/src/commands/identity/mod.rs
@@ -11,37 +11,19 @@ pub(crate) mod new;
 pub(crate) mod principal;
 pub(crate) mod rename;
 
+/// Manage your identities
 #[derive(Debug, Subcommand)]
 pub(crate) enum Command {
-    /// Display the ICP ledger account identifier for the current identity
     AccountId(account_id::AccountIdArgs),
-
-    /// Display the currently selected identity
     Default(default::DefaultArgs),
-
-    /// Delete an identity
     Delete(delete::DeleteArgs),
-
-    /// Print the PEM file for the identity
     Export(export::ExportArgs),
-
-    /// Import a new identity
     Import(import::ImportArgs),
-
-    /// Link an external key to a new identity
     #[command(subcommand)]
     Link(link::Command),
-
-    /// List the identities
     List(list::ListArgs),
-
-    /// Create a new identity
     New(new::NewArgs),
-
-    /// Display the principal for the current identity
     Principal(principal::PrincipalArgs),
-
-    /// Rename an identity
     Rename(rename::RenameArgs),
 }