+++ crates/icp/src/lib.rs
@@ -368,6 +368,7 @@ impl MockProjectLoader {
                         subnets: None,
                         bitcoind_addr: None,
                         dogecoind_addr: None,
+                        version: None,
                     })),
                 },
             },
@@ -388,6 +389,7 @@ impl MockProjectLoader {
                         subnets: None,
                         bitcoind_addr: None,
                         dogecoind_addr: None,
+                        version: None,
                     })),
                 },
             },
@@ -397,7 +399,8 @@ impl MockProjectLoader {
             name: "ic".to_string(),
             configuration: Configuration::Connected {
                 connected: Connected {
-                    url: "https://ic0.app".to_string(),
+                    api_url: "https://icp-api.io".parse().unwrap(),
+                    http_gateway_url: Some("https://icp0.io".parse().unwrap()),
                     root_key: None,
                 },
             },