+++ crates/icp/src/network/mod.rs
@@ -8,12 +8,13 @@ use snafu::prelude::*;
 pub use directory::{LoadPidError, NetworkDirectory, SavePidError};
 pub use managed::run::{RunNetworkError, run_network};
 use strum::EnumString;
+use url::Url;
 
 use crate::{
     CACHE_DIR, ICP_BASE, Network,
     manifest::{
         ProjectRootLocate, ProjectRootLocateError,
-        network::{Connected as ManifestConnected, Gateway as ManifestGateway, Mode},
+        network::{Connected as ManifestConnected, Endpoints, Gateway as ManifestGateway, Mode},
     },
     network::access::{
         GetNetworkAccessError, NetworkAccess, get_connected_network_access,
@@ -92,6 +93,7 @@ pub struct ManagedLauncherConfig {
     pub subnets: Option<Vec<SubnetKind>>,
     pub bitcoind_addr: Option<Vec<String>>,
     pub dogecoind_addr: Option<Vec<String>>,
+    pub version: Option<String>,
 }
 
 #[derive(
@@ -132,6 +134,7 @@ impl ManagedMode {
             subnets: None,
             bitcoind_addr: None,
             dogecoind_addr: None,
+            version: None,
         }))
     }
 }
@@ -155,8 +158,11 @@ pub struct ManagedImageConfig {
 #[derive(Clone, Debug, Deserialize, PartialEq, JsonSchema, Serialize)]
 #[serde(rename_all = "kebab-case")]
 pub struct Connected {
-    /// The URL this network can be reached at.
-    pub url: String,
+    /// The URL this network's API can be reached at.
+    pub api_url: Url,
+
+    /// The URL this network's HTTP gateway can be reached at.
+    pub http_gateway_url: Option<Url>,
 
     /// The root key of this network
     pub root_key: Option<Vec<u8>>,
@@ -203,9 +209,21 @@ impl From<ManifestGateway> for Gateway {
 
 impl From<ManifestConnected> for Connected {
     fn from(value: ManifestConnected) -> Self {
-        let url = value.url.clone();
-        let root_key = value.root_key.map(|rk| rk.0);
-        Connected { url, root_key }
+        match value.endpoints {
+            Endpoints::Implicit { url } => Connected {
+                api_url: url.clone(),
+                http_gateway_url: Some(url),
+                root_key: value.root_key.map(|rk| rk.0),
+            },
+            Endpoints::Explicit {
+                api_url,
+                http_gateway_url,
+            } => Connected {
+                api_url,
+                http_gateway_url,
+                root_key: value.root_key.map(|rk| rk.0),
+            },
+        }
     }
 }
 
@@ -221,10 +239,21 @@ impl From<Mode> for Configuration {
                     subnets,
                     bitcoind_addr,
                     dogecoind_addr,
+                    version,
                 } => {
                     let gateway: Gateway = match gateway {
                         Some(g) => g.into(),
                         None => Gateway::default(),
+                    };
+                    let version = match version {
+                        Some(v) => {
+                            if v.starts_with('v') {
+                                Some(v)
+                            } else {
+                                Some(format!("v{v}"))
+                            }
+                        }
+                        None => None,
                     };
                     Configuration::Managed {
                         managed: Managed {
@@ -236,6 +265,7 @@ impl From<Mode> for Configuration {
                                 subnets,
                                 bitcoind_addr,
                                 dogecoind_addr,
+                                version,
                             })),
                         },
                     }
@@ -408,6 +438,7 @@ mod tests {
                 subnets: None,
                 bitcoind_addr: Some(vec!["127.0.0.1:18444".to_string()]),
                 dogecoind_addr: None,
+                version: None,
             }),
         });